requirements.txt
cmsplugin-photologue
django-photologue
